+++ one/tests/test_params.py
@@ -87,20 +87,6 @@ class TestONEParamUtil(unittest.TestCase):
         self.assertIsInstance(path, Path)
         self.assertEqual('path/to/params/.one', path.as_posix())
 
-    def test_get_rest_dir(self):
-        """Test for one.params.get_rest_dir"""
-        par_dir = Path('path', 'to', 'params')
-        url = 'https://sub.domain.net/'
-        with mock.patch('iblutil.io.params.getfile', new=partial(util.get_file, par_dir)):
-            path1 = params.get_rest_dir()
-            path2 = params.get_rest_dir(url)
-
-        expected = ('path', 'to', 'params', '.one', '.rest')
-        self.assertCountEqual(expected, path1.parts)
-
-        expected = (*expected, 'sub.domain.net', 'https')
-        self.assertCountEqual(expected, path2.parts)
-
     def test_get_default_client(self):
         """Test for one.params.get_default_client"""
         temp_dir = util.set_up_env()